photography V4.1.1

+++ wp-content/themes/photography/lib/admin.action.lib.php
@@ -1,6 +1,6 @@
 <?php
 //If clear cache
-if(is_admin() && isset($_POST['method']) && !empty($_POST['method']) && $_POST['method'] == 'clear_cache')
+if(is_admin() && current_user_can('manage_options') && isset($_POST['method']) && !empty($_POST['method']) && $_POST['method'] == 'clear_cache')
 {
 	//Get theme cache folder
 	$upload_dir = wp_upload_dir();
@@ -26,7 +26,7 @@ if(is_admin() && isset($_POST['method']) && !empty($_POST['method']) && $_POST['
 }
 
 //If export settings
-if(is_admin() && isset($_POST['pp_export_current']) && !empty($_POST['pp_export_current']))
+if(is_admin() && current_user_can('manage_options') && isset($_POST['pp_export_current']) && !empty($_POST['pp_export_current']))
 {
 	$json_file_name = THEMENAME.'Theme_Export_'.date('m-d-Y_hia');
 
@@ -57,7 +57,7 @@ if(is_admin() && isset($_POST['pp_export_current']) && !empty($_POST['pp_export_
 }
 
 //If delete sidebar
-if(is_admin() && isset($_POST['sidebar_id']) && !empty($_POST['sidebar_id']))
+if(is_admin() && current_user_can('manage_options') && isset($_POST['sidebar_id']) && !empty($_POST['sidebar_id']))
 {
 	$current_sidebar = get_option('pp_sidebar');
 	
@@ -72,7 +72,7 @@ if(is_admin() && isset($_POST['sidebar_id']) && !empty($_POST['sidebar_id']))
 }
 
 //If delete image
-if(is_admin() && isset($_POST['field_id']) && !empty($_POST['field_id']) && isset($_GET["page"]) && $_GET["page"] == "functions.php" )
+if(is_admin() && current_user_can('manage_options') && isset($_POST['field_id']) && !empty($_POST['field_id']) && isset($_GET["page"]) && $_GET["page"] == "functions.php" )
 {
 	$current_val = get_option($_POST['field_id']);
 	delete_option( $_POST['field_id'] );
